+++ appseed/openssl/ssl/ssl_rsa.c
@@ -751,7 +751,7 @@ int SSL_CTX_use_certificate_chain_file(SSL_CTX *ctx, const char *file)
 		while ((ca = PEM_read_bio_X509(in,NULL,ctx->default_passwd_callback,ctx->default_passwd_callback_userdata))
 			!= NULL)
 			{
-			r = SSL_CTX_add_extra_chain_cert(ctx, ca);
+			r = (int) SSL_CTX_add_extra_chain_cert(ctx, ca);
 			if (!r) 
 				{
 				X509_free(ca);
